+++ yolo_v3/yolo_v3/src/network.c
@@ -92,12 +92,12 @@ float get_current_rate(network *net)
     size_t batch_num = get_current_batch(net);
     int i;
     float rate;
-    if (batch_num < net->burn_in) return net->learning_rate * (float)pow((float)batch_num / net->burn_in, net->power);
+    if (batch_num < net->burn_in) return net->learning_rate * powf((float)batch_num / net->burn_in, net->power);
     switch (net->policy) {
         case CONSTANT:
             return net->learning_rate;
         case STEP:
-            return net->learning_rate * pow(net->scale, batch_num/net->step);
+            return net->learning_rate * powf(net->scale, batch_num/net->step);
         case STEPS:
             rate = net->learning_rate;
             for(i = 0; i < net->num_steps; ++i){
@@ -106,13 +106,13 @@ float get_current_rate(network *net)
             }
             return rate;
         case EXP:
-            return net->learning_rate * (float)pow(net->gamma, batch_num);
+            return net->learning_rate * powf(net->gamma, batch_num);
         case POLY:
-            return net->learning_rate * (float)pow(1 - (float)batch_num / net->max_batches, net->power);
+            return net->learning_rate * powf(1 - (float)batch_num / net->max_batches, net->power);
         case RANDOM:
-            return net->learning_rate * (float)pow(rand_uniform(0,1), net->power);
+            return net->learning_rate * powf(rand_uniform(0,1), net->power);
         case SIG:
-            return net->learning_rate * (1./(1.+exp(net->gamma*(batch_num - net->step))));
+            return net->learning_rate * (1.f/(1.f+expf(net->gamma*(batch_num - net->step))));
         default:
             fprintf(stderr, "Policy is weird!\n");
             return net->learning_rate;
@@ -673,8 +673,8 @@ void compare_networks(network *n1, network *n2, data test)
         }
     }
     printf("%5d %5d\n%5d %5d\n", a, b, c, d);
-    float num = (float)pow((abs(b - c) - 1.f), 2.f);
-    float den = b + c;
+    float num = powf((abs(b - c) - 1.f), 2.f);
+    float den = (float)(b + c);
     printf("%f\n", num/den); 
 }
